+++ twitchirc/twitchirc/bot.py
@@ -13,7 +13,6 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import atexit
 import sched
 import select
 import typing
@@ -23,15 +22,52 @@ import twitchirc  # Import self.
 
 class Bot(twitchirc.Connection):
     def schedule_event(self, delay, priority, function, args: tuple, kwargs: dict):
+        """
+        Schedule an event using the scheduler. Wrapper for :py:meth:`shed.scheduler.enter`
+
+        :param delay: Delay the task will be executed in.
+        :param priority: Task's priority.
+        :param function: Function that will be run.
+        :param args: Arguments to that function.
+        :param kwargs: Keyword arguments to that function.
+
+        :return: Event object.
+        """
         return self.scheduler.enter(delay, priority, function, args, kwargs)
 
     def schedule_event_absolute(self, time, priority, function, args: tuple, kwargs: dict):
+        """
+        Schedule an event using the scheduler with a time, not a delay. Wrapper for :py:meth:`shed.scheduler.enterabs`
+
+        :param time: Time the task will be executed.
+        :param priority: Task's priority.
+        :param function: Function that will be run.
+        :param args: Arguments to that function.
+        :param kwargs: Keyword arguments to that function.
+
+        :return: Event object.
+        """
         return self.scheduler.enterabs(time, priority, function, args, kwargs)
 
     def schedule_repeated_event(self, delay, priority, function, args: tuple, kwargs: dict):
+        """
+        Schedule a repeated event using the scheduler.
+
+        :param delay: Delay the task will be executed in.
+        :param priority: Task's priority.
+        :param function: Function that will be run.
+        :param args: Arguments to that function.
+        :param kwargs: Keyword arguments to that function.
+
+        :return: Event object.
+
+        NOTE: There is currently no way to cancel an event. You have to do it manually.
+        """
+
         def run_event():
             function()
             self.scheduler.enter(delay, priority, run_event, args, kwargs)
+
         return self.scheduler.enter(delay, priority, run_event, args, kwargs)
 
     def run_commands_from_file(self, file_object):
@@ -60,18 +96,21 @@ class Bot(twitchirc.Connection):
         self._in_rc_mode = False
 
     def __init__(self, address, username: str, password: typing.Union[str, None] = None, port: int = 6667,
-                 no_connect=False, storage=None, no_atexit=False):
+                 no_connect=False, storage=None, no_atexit=False, secure=False, message_cooldown=3):
         """
         A bot class.
 
         :param address: Address to connect to
         :param username: Username to use
-        :param password: Password if needed.
+        :param password: Password if needed, otherwise None can be used.
         :param port: Irc port.
-        :param no_connect: Don't connect to the chat straight away
-        :param storage: twitchirc.Storage compatible object to use as the storage.
+        :param no_connect: Don't connect to the chat straight away.
+        :param no_atexit: Don't use atexit to automatically disconnect.
+        :param storage: A :py:class:`twitchirc.Storage`.
+        :param secure: Use a secure connection, SSL.
+        :param message_cooldown: Per-channel cooldown for messages.
         """
-        super().__init__(address, port, no_atexit=True)
+        super().__init__(address, port, no_atexit=no_atexit, message_cooldown=message_cooldown, secure=secure)
         self.scheduler = sched.scheduler()
         self._in_rc_mode = False
         if not no_connect:
@@ -97,21 +136,38 @@ class Bot(twitchirc.Connection):
 
         Available handlers:
 
-        * pre_disconnect, args: ()
-        * post_disconnect, args: ()
-        * pre_save, args: ()
-        * post_save, args: ()
-        * start, args: ()
-        * permission_error, args: (message, command, missing_permissions)
-          -> message the ChannelMessage during which this permission_error was triggered.
-          -> command the Command object that triggered it.
-             WARN: `command` can be None if `check_permissions` was called (not `check_permissions_from_command`).
-          -> missing_permissions permissions that are missing to run the command.
-        * any_msg, args: (message)
-        * chat_msg, args: (message)
+        +----------------+----------------------------------------+---------------------------------------------------+
+        |name            |arguments                               |explanation                                        |
+        +================+========================================+===================================================+
+        |pre_disconnect  |()                                      |                                                   |
+        +----------------+----------------------------------------+---------------------------------------------------+
+        |post_disconnect |()                                      |                                                   |
+        +----------------+----------------------------------------+---------------------------------------------------+
+        |pre_save        |()                                      |                                                   |
+        +----------------+----------------------------------------+---------------------------------------------------+
+        |post_save       |()                                      |                                                   |
+        +----------------+----------------------------------------+---------------------------------------------------+
+        |start           |()                                      |                                                   |
+        +----------------+----------------------------------------+---------------------------------------------------+
+        |permission_error|(message, command, missing_permissions) | message -- the ChannelMessage during which this   |
+        |                |                                        | permission_error was triggered.                   |
+        |                |                                        |                                                   |
+        |                |                                        | command -- the Command object that triggered it.  |
+        |                |                                        |                                                   |
+        |                |                                        | WARN: `command` can be None if `check_permissions`|
+        |                |                                        | was called (not `check_permissions_from_command`).|
+        |                |                                        |                                                   |
+        |                |                                        | missing_permissions -- permissions that are       |
+        |                |                                        | missing to run the command.                       |
+        +----------------+----------------------------------------+---------------------------------------------------+
+        |any_msg         |(message)                               |                                                   |
+        +----------------+----------------------------------------+---------------------------------------------------+
+        |chat_msg        |(message)                               |                                                   |
+        +----------------+----------------------------------------+---------------------------------------------------+        
         """
 
         self.prefix = '!'
+        """Bot prefix. Can be set to any string. Recommended to set before registering any commands"""
         self.storage = storage
         self.on_unknown_command = 'ignore'
         """
@@ -125,18 +181,11 @@ class Bot(twitchirc.Connection):
         * chat_message - send a chat message saying that the command is invalid.
         """
         self.permissions = twitchirc.PermissionList()
-        if no_atexit:
-            @atexit.register
-            def close_self():
-                try:
-                    self.stop()
-                except:
-                    pass
 
     def add_command(self, command: str,
                     forced_prefix: typing.Optional[str] = None,
                     enable_local_bypass: bool = True,
-                    required_permissions: typing.List[str] = []) \
+                    required_permissions: typing.Optional[typing.List[str]] = None) \
             -> typing.Callable[[typing.Callable[[twitchirc.ChannelMessage],
                                                 typing.Any]], twitchirc.Command]:
         # I'm sorry if you are reading this definition
@@ -147,13 +196,17 @@ class Bot(twitchirc.Connection):
         This function is a decorator.
 
         :param command: Command to be registered.
-        :param forced_prefix: Force a prefix to on this command. This is useful when you can change the prefix in the
-        bot.
-        :param enable_local_bypass: If False this function will ignore the permissions
+        :param forced_prefix: Force a prefix to on this command. This is useful when you can change the prefix in \
+        the bot.
+        :param enable_local_bypass: If False this function will ignore the permissions \
         `twitchirc.bypass.permission.local.*`. This is useful when creating a command that can change global settings.
         :param required_permissions: Permissions required to run this command.
-        :return: The `function` parameter. This is for using this as a decorator.
+
+        This is a decorator.
         """
+
+        if required_permissions is None:
+            required_permissions = []
 
         def decorator(func: typing.Callable) -> twitchirc.Command:
             cmd = twitchirc.Command(chat_command=command,
@@ -161,6 +214,7 @@ class Bot(twitchirc.Connection):
                                     enable_local_bypass=enable_local_bypass)
             cmd.permissions_required.extend(required_permissions)
             self.commands.append(cmd)
+            self.call_middleware('add_command', (cmd,), cancelable=False)
             return cmd
 
         return decorator
@@ -172,12 +226,20 @@ class Bot(twitchirc.Connection):
 
         :param message: Message received.
         :param permissions: Permissions required.
-        :param enable_local_bypass: If False this function will ignore the permissions
+        :param enable_local_bypass: If False this function will ignore the permissions \
         `twitchirc.bypass.permission.local.*`. This is useful when creating a command that can change global settings.
         :return: A list of missing permissions.
 
         NOTE `permission_error` handlers are called if this function would return a non empty list.
         """
+        o = self.call_middleware('permission_check', dict(user=message.user, permissions=permissions,
+                                                          message=message), cancelable=True)
+        if o is False:
+            return ['impossible.event_canceled']
+
+        if isinstance(o, tuple):
+            return o[1]
+
         missing_permissions = []
         if message.user not in self.permissions:
             missing_permissions = permissions
@@ -205,6 +267,14 @@ class Bot(twitchirc.Connection):
 
         NOTE `permission_error` handlers are called if this function would return a non empty list.
         """
+        o = self.call_middleware('permission_check', dict(user=message.user, permissions=command.permissions_required,
+                                                          message=message, command=command), cancelable=True)
+        if o is False:
+            return ['impossible.event_canceled']
+
+        if isinstance(o, tuple):
+            return o[1]
+
         missing_permissions = []
         if message.user not in self.permissions:
             missing_permissions = command.permissions_required
@@ -224,6 +294,7 @@ class Bot(twitchirc.Connection):
         return missing_permissions
 
     def _call_command_handlers(self, message: twitchirc.ChannelMessage):
+        """Handle commands."""
         if message.text.startswith(self.prefix):
             was_handled = False
             if ' ' not in message.text:
@@ -241,6 +312,7 @@ class Bot(twitchirc.Connection):
             self._call_forced_prefix_commands(message)
 
     def _call_forced_prefix_commands(self, message):
+        """Handle commands with forced prefixes."""
         for handler in self.commands:
             if handler.forced_prefix is None:
                 continue
@@ -248,6 +320,7 @@ class Bot(twitchirc.Connection):
                 handler(message)
 
     def _do_unknown_command(self, message):
+        """Handle unknown commands."""
         if self.on_unknown_command == 'warn':
             twitchirc.warn(f'Unknown command {message!r}')
         elif self.on_unknown_command == 'chat_message':
@@ -275,10 +348,16 @@ class Bot(twitchirc.Connection):
             return
 
     def _select_socket(self):
+        """Check if the socket has any data to receive."""
         sel_output = select.select([self.socket], [], [], 0.1)
         return bool(sel_output[0])
 
     def _run_once(self):
+        """
+        Do everything needed to run, but don't loop. This can be used as a non-blocking version of :py:meth:`run`.
+
+        :return: False if the bot should quit, True if everything went right, 'RECONNECT' if the bot needs to reconnect.
+        """
         if self.socket is None:  # self.disconnect() was called.
             return False
         if not self._select_socket():  # no data in socket, assume all messages where handled last time and return
@@ -309,6 +388,11 @@ class Bot(twitchirc.Connection):
         return True
 
     def _run(self):
+        """
+        Brains behind :py:meth:`run`. Doesn't include the `KeyboardInterrupt` handler.
+
+        :return: nothing.
+        """
         if self.socket is None:
             self.connect(self.username, self._password)
         self.hold_send = False
@@ -366,12 +450,12 @@ class Bot(twitchirc.Connection):
         Send a message to the server.
 
         :param message: message to send
-        :param queue: Queue for the message to be in. This will be automaticcally overriden if the message is a
-        ChannelMessage. It will be set to `message.channel`.
+        :param queue: Queue for the message to be in. This will be automatically overridden if the message is a \
+        ChannelMessage. It will be set to :py:meth:`ChannelMessage.channel`.
 
         :return: nothing
 
-        NOTE The message will not be sent instantely and this is intended. If you would send lots of messages Twitch
+        NOTE The message will not be sent instantly and this is intended. If you would send lots of messages Twitch
         will not forward any of them to the chat clients.
         """
         if self._in_rc_mode: